feat: add delete all messages card

+++ src/pages/api/delete-all-messages.ts
@@ -1,4 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { PIN_CODE } from "src/data/constants";
 import { middleware } from "src/lib/middleware";
 import { deleteAllMessages } from "src/utils/firebaseMessageHandler";
 
@@ -8,8 +9,12 @@ export default middleware(async function handler(
 ) {
   try {
     if (req.method === "DELETE") {
-      await deleteAllMessages();
-      return res.json({ message: "nice", success: true });
+      const { pin_code } = req.body;
+
+      if (pin_code === PIN_CODE) {
+        await deleteAllMessages();
+        return res.json({ message: "nice", success: true });
+      }
     }
   } catch (e) {
     console.error(e);
